exp/services/webauth: rename commands package to cmd (#2278)
### What

Rename the `exp/services/webauth/internal/commands` package to `exp/services/webauth/cmd`.

### Why

The cobra commands for the `webauth` app live in the `internal/commands` package but we have a couple apps that are using the convention to put commands in a top-level `cmd` package. E.g. Horizon, Ticker. This makes `webauth` more consistent with those applications.

This idea was discussed here: https://github.com/stellar/go/pull/2267#discussion_r379692099
